Start with chat rooll

+++ modules/handlers/users_handlers/likes.py
@@ -9,7 +9,7 @@ from modules.keyboards import user_likes_kb, likes_kb, likes_in_profile_kb
 from modules.sql_func import update_db, read_by_name, search_person, join_profile_all, read_all_2, \
     insert_likes_presents_db
 from modules.handlers.handlers_func import edit_text_call
-from modules.functions.simple_funcs import update_age_period
+from modules.functions.simple_funcs import update_age_period, get_right_left_btn
 
 
 # Profile menu
@@ -49,7 +49,6 @@ async def start_menu(call: types.CallbackQuery):
 @dp.callback_query_handler(state=UserLikes.start, text='user_likes')
 async def start_menu(call: types.CallbackQuery):
     all_likes = read_by_name(name='from_tg_id', id_data=call.from_user.id, table='likes')
-    update_db(table='fast_info', name='fast_1', data=str(all_likes), id_data=call.from_user.id)
     if str(all_likes) == '[]':
         return await call.message.answer('Увы, но у вас еще нет взаимных симпатий, ставьте лайки другим '
                                          'пользователям и вы обязательно найдете взаимную симпатию 😇')
@@ -61,7 +60,6 @@ async def start_menu(call: types.CallbackQuery):
 @dp.callback_query_handler(state=UserLikes.start, text='user_presents_send')
 async def start_menu(call: types.CallbackQuery):
     all_likes = read_by_name(name='tg_id', id_name="from_tg_id", id_data=call.from_user.id, table='presents')
-    update_db(table='fast_info', name='fast_1', data=str(all_likes), id_data=call.from_user.id)
     if str(all_likes) == '[]':
         return await call.message.answer('Вы еще никому не отправляли 🎁 подарки.')
 
@@ -72,7 +70,7 @@ async def start_menu(call: types.CallbackQuery):
 @dp.callback_query_handler(state=UserLikes.start, text='user_presents_from')
 async def start_menu(call: types.CallbackQuery):
     all_likes = read_by_name(name='from_tg_id', id_name="tg_id", id_data=call.from_user.id, table='presents')
-    update_db(table='fast_info', name='fast_1', data=str(all_likes), id_data=call.from_user.id)
+
     if str(all_likes) == '[]':
         return await call.message.answer('😔 Увы, но вы еще не получили ни одно 🎁 подарка')
 
@@ -83,31 +81,49 @@ async def start_menu(call: types.CallbackQuery):
 @dp.callback_query_handler(state=UserLikes.likes_for_me, text_contains='like_kb_')
 async def start_menu(call: types.CallbackQuery):
     user_id = call.data.split('_')[2]
+    if str(user_id) == 'stop':
+        return await bot.answer_inline_query(call.id)
+
+    all_likes = read_by_name(name='from_tg_id', id_data=call.from_user.id, table='likes')
+
+    index, left, right = get_right_left_btn(check_id=user_id, all_likes=all_likes)
+
     text, photo_id = show_other_profile(user_id=int(user_id), user_finder_id=call.from_user.id)
     await call.message.delete()
-    text = f'👍 Вас лайкнули <b>[1/2]</b>\n\n{text}'
+    text = f'👍 Вас лайкнули <b>[{index}/{len(all_likes)}]</b>\n\n{text}'
     await bot.send_photo(caption=text, photo=photo_id, chat_id=call.from_user.id, parse_mode='html',
-                         reply_markup=likes_in_profile_kb())
+                         reply_markup=likes_in_profile_kb(left=left, right=right))
     await UserLikes.likes_for_me.set()
 
 
 @dp.callback_query_handler(state=UserLikes.presents_from_me, text_contains='like_kb_')
 async def start_menu(call: types.CallbackQuery):
     user_id = call.data.split('_')[2]
+    if str(user_id) == 'stop':
+        return await bot.answer_inline_query(call.id)
+
+    all_likes = read_by_name(name='tg_id', id_name="from_tg_id", id_data=call.from_user.id, table='presents')
+    index, left, right = get_right_left_btn(check_id=user_id, all_likes=all_likes)
+
     text, photo_id = show_other_profile(user_id=int(user_id), user_finder_id=call.from_user.id)
     await call.message.delete()
-    text = f'🎁 Кому отправил <b>[1/2]</b>\n\n{text}'
+    text = f'🎁 Кому отправил <b>[{index}/{len(all_likes)}]</b>\n\n{text}'
     await bot.send_photo(caption=text, photo=photo_id, chat_id=call.from_user.id, parse_mode='html',
-                         reply_markup=likes_in_profile_kb())
+                         reply_markup=likes_in_profile_kb(left=left, right=right))
     await UserLikes.presents_from_me.set()
 
 
 @dp.callback_query_handler(state=UserLikes.presents_for_me, text_contains='like_kb_')
 async def start_menu(call: types.CallbackQuery):
     user_id = call.data.split('_')[2]
+    if str(user_id) == 'stop':
+        return await bot.answer_inline_query(call.id)
+    all_likes = read_by_name(name='from_tg_id', id_name="tg_id", id_data=call.from_user.id, table='presents')
+    index, left, right = get_right_left_btn(check_id=user_id, all_likes=all_likes)
+
     text, photo_id = show_other_profile(user_id=int(user_id), user_finder_id=call.from_user.id)
     await call.message.delete()
-    text = f'🎁 От кого получил <b>[1/2]</b>\n\n{text}'
+    text = f'🎁 От кого получил <b>[{index}/{len(all_likes)}]</b>\n\n{text}'
     await bot.send_photo(caption=text, photo=photo_id, chat_id=call.from_user.id, parse_mode='html',
-                         reply_markup=likes_in_profile_kb())
+                         reply_markup=likes_in_profile_kb(left=left, right=right))
     await UserLikes.presents_for_me.set()
